+++ lib/mix/tasks/compile/boundary.ex
@@ -209,8 +209,8 @@ defmodule Mix.Tasks.Compile.Boundary do
     )
   end
 
-  defp to_diagnostic_error({:ignored_dep, dep}) do
-    diagnostic("ignored boundary #{inspect(dep.name)} is listed as a dependency",
+  defp to_diagnostic_error({:check_in_false_dep, dep}) do
+    diagnostic("boundary #{inspect(dep.name)} can't be a dependency because it has check.in set to false",
       file: Path.relative_to_cwd(dep.file),
       position: dep.line
     )
@@ -282,6 +282,14 @@ defmodule Mix.Tasks.Compile.Boundary do
     diagnostic(message, file: Path.relative_to_cwd(error.file), position: error.line)
   end
 
+  defp to_diagnostic_error({:unknown_option, %{name: :ignore?, value: value} = data}) do
+    diagnostic(
+      "ignore?: #{value} is deprecated, use check: [in: #{not value}, out: #{not value}] instead",
+      file: Path.relative_to_cwd(data.file),
+      position: data.line
+    )
+  end
+
   defp to_diagnostic_error({:unknown_option, data}) do
     diagnostic("unknown option #{inspect(data.name)}",
       file: Path.relative_to_cwd(data.file),
@@ -289,15 +297,22 @@ defmodule Mix.Tasks.Compile.Boundary do
     )
   end
 
-  defp to_diagnostic_error({:dep_in_ignored_boundary, data}) do
-    diagnostic("deps can't be provided in an ignored boundary",
+  defp to_diagnostic_error({:deps_in_check_out_false, data}) do
+    diagnostic("deps can't be listed if check.out is set to false",
       file: Path.relative_to_cwd(data.file),
       position: data.line
     )
   end
 
-  defp to_diagnostic_error({:export_in_ignored_boundary, data}) do
-    diagnostic("exports can't be provided in an ignored boundary",
+  defp to_diagnostic_error({:apps_in_check_out_false, data}) do
+    diagnostic("check apps can't be listed if check.out is set to false",
+      file: Path.relative_to_cwd(data.file),
+      position: data.line
+    )
+  end
+
+  defp to_diagnostic_error({:exports_in_check_in_false, data}) do
+    diagnostic("can't export modules if check.in is set to false",
       file: Path.relative_to_cwd(data.file),
       position: data.line
     )
@@ -307,6 +322,20 @@ defmodule Mix.Tasks.Compile.Boundary do
     diagnostic("invalid type",
       file: Path.relative_to_cwd(data.file),
       position: data.line
+    )
+  end
+
+  defp to_diagnostic_error({:invalid_ignores, boundary}) do
+    diagnostic("can't disable checks in a sub-boundary",
+      file: Path.relative_to_cwd(boundary.file),
+      position: boundary.line
+    )
+  end
+
+  defp to_diagnostic_error({:ancestor_with_ignored_checks, boundary, ancestor}) do
+    diagnostic("sub-boundary inside a boundary with disabled checks (#{inspect(ancestor.name)})",
+      file: Path.relative_to_cwd(boundary.file),
+      position: boundary.line
     )
   end